--- yumyum_front/src/components/Main.tsx
@@ -1,7 +1,0 @@
-import React from 'react'
-
-export default function Main() {
-  return (
-    <div>Main</div>
-  )
-}
